refactor: move parsers to separate files
