+++ src/parsers/__tests__/css.test.ts
@@ -1,0 +1,7 @@
+import { parseCss } from "../css";
+
+describe("parseCss()", () => {
+  test("throws when passed invalid css", () => {
+    expect(() => parseCss("color #fff")).toThrow();
+  });
+});
