+++ src/parsers/__tests__/js.test.ts
@@ -1,0 +1,27 @@
+import { parseJsObject } from "../js";
+
+describe("parseJsObject()", () => {
+  test("parses an object expression", () => {
+    const [result] = parseJsObject(`{ myKey: "myValue" }`);
+    expect(result.type).toBe("ObjectExpression");
+  });
+
+  test("wraps input in braces if needed", () => {
+    const [, rawLines] = parseJsObject(`someKey: "someValue"`);
+    expect(rawLines).toBe(`{someKey: "someValue"}`);
+  });
+
+  test("throws when passed an expression that's not an object", () => {
+    const callWithBadParam = () => {
+      parseJsObject("[notAnObject]");
+    };
+    expect(callWithBadParam).toThrow();
+  });
+
+  test("throws when passed something that's not an expression", () => {
+    const callWithBadParam = () => {
+      parseJsObject("let x = 1");
+    };
+    expect(callWithBadParam).toThrow();
+  });
+});
