+++ src/parsers/js.ts
@@ -1,8 +1,7 @@
 import * as babelParser from "@babel/parser";
 import { codeFrameColumns } from "@babel/code-frame";
 import { Expression, Node, ObjectExpression } from "@babel/types";
-import { getErrorLocation } from "./getErrorLocation";
-import { ChildNode, Declaration, parse, Root } from "postcss";
+import { getErrorLocation } from "../utils";
 
 export function parseJsObject(input: string): [ObjectExpression, string] {
   let rawLines: string;
@@ -50,42 +49,4 @@ export function nodeToString(node: Node, code: string): string {
 
   // Copy the string value directly from the input code
   return code.slice(start, end).trim();
-}
-
-/**
- * Finds all css delcarations
- * @param root PostCss AST Root
- */
-function getCssProps(root: Root): Declaration[] {
-  const { nodes } = root;
-
-  if (!nodes) {
-    return [];
-  }
-
-  function recurse(nodes: ChildNode[], declarations: Declaration[] = []) {
-    nodes.forEach((item) => {
-      if (item.type === "decl") {
-        declarations.push(item);
-      } else if (item.type === "rule" && item?.nodes) {
-        recurse(item.nodes, declarations);
-      }
-    });
-
-    return declarations;
-  }
-
-  return recurse(nodes);
-}
-/**
- * Parses css into PostCss AST
- * @param css css string
- */
-export function parseCss(css: string): Declaration[] {
-  try {
-    const root: Root = parse(css);
-    return getCssProps(root);
-  } catch (e) {
-    throw new Error(`${e.reason} on line ${e.line}`);
-  }
 }
