+++ src/utils/index.ts
@@ -1,4 +1,4 @@
 export * from "./defaultHightlightProps";
+export * from "./getErrorLocation";
 export * from "./exampleCode";
-export * from "./parser";
 export * from "./usePrevious";
